requirements.txt
discord.py==1.7.3
dislash.py==1.4.9